+++ src/obelist/cli/cmds/__init__.py
@@ -1,13 +1,12 @@
 import click
 
-from obelist import cli
-from obelist.core.app import Application
-
 from . import debug
 from . import format
 from . import list
 from . import parse
 from . import path
+from obelist import cli
+from obelist.core.app import Application
 
 
 @click.group(cls=cli.StylizedGroup)